requirements.txt
# Core framework
fastapi>=0.103.1,<0.104.0
uvicorn[standard]>=0.23.2,<0.24.0
pydantic>=2.3.0,<3.0.0
starlette>=0.27.0,<0.28.0

# Authentication
pyjwt>=2.8.0,<3.0.0
passlib>=1.7.4,<2.0.0
python-jose>=3.3.0,<4.0.0
python-multipart>=0.0.6,<0.0.7

# Database
sqlalchemy>=2.0.20,<3.0.0
databases>=0.9.0,<0.10.0
asyncpg>=0.28.0,<0.29.0
alembic>=1.12.0,<2.0.0

# ML & NLP libraries
onnxruntime>=1.15.1,<2.0.0
torch>=2.0.1,<3.0.0
torchvision>=0.15.2,<0.16.0
torchaudio>=2.0.2,<2.1.0
transformers==4.37.2
tokenizers>=0.14.0,<0.19.0

sentencepiece>=0.1.99,<0.2.0
sacremoses>=0.0.53,<0.1.0
numpy>=1.25.2,<2.0.0
scipy>=1.11.2,<2.0.0
scikit-learn>=1.3.0,<2.0.0
huggingface_hub==0.15.1

# Vector databases/embeddings
faiss-cpu>=1.7.4,<2.0.0
chromadb>=0.4.15,<0.5.0
hnswlib>=0.7.0,<0.8.0

# Data processing
pandas>=2.1.0,<3.0.0
pyarrow>=13.0.0,<14.0.0

# Document processing
docx2txt>=0.8,<0.9
fasttext>=0.9.2,<0.10.0
pypdf>=3.16.0,<4.0.0
python-docx>=0.8.11,<0.9.0
pymupdf>=1.22.5,<2.0.0
pytesseract>=0.3.10,<0.4.0
pillow>=10.0.1,<11.0.0
langdetect>=1.0.9,<2.0.0
chardet>=5.2.0,<6.0.0

# Utilities
colorama
psutil>=5.9.5,<6.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.24.1,<0.25.0
pydantic-settings>=2.0.3,<3.0.0
tenacity>=8.2.3,<9.0.0
tqdm>=4.66.1,<5.0.0
aiofiles>=23.2.1,<24.0.0
ujson>=5.8.0,<6.0.0
orjson>=3.9.5,<4.0.0
msgpack>=1.0.5,<2.0.0

# Monitoring and logging
prometheus-client>=0.17.1,<0.18.0
structlog>=23.1.0,<24.0.0
python-json-logger>=2.0.7,<3.0.0
opencensus>=0.11.2,<0.12.0
opencensus-ext-azure>=1.1.9,<2.0.0
statsd>=4.0.1,<5.0.0

# Async-related
aiohttp>=3.8.5,<4.0.0
aiokafka>=0.8.1,<0.9.0
websockets>=11.0.3,<12.0.0

# API-related
requests>=2.31.0,<3.0.0
itsdangerous>=2.1.2,<3.0.0
jinja2>=3.1.2,<4.0.0
email-validator>=2.0.0,<3.0.0

# RAG-specific
llama-index>=0.8.22,<0.9.0
langchain>=0.0.286,<0.1.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.1,<0.22.0
pytest-cov>=4.1.0,<5.0.0
hypothesis>=6.86.0,<7.0.0
pytest-xdist>=3.3.1,<4.0.0
faker>=19.6.1,<20.0.0

# Development tools
black>=23.7.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.5.1,<2.0.0
flake8>=6.1.0,<7.0.0
pre-commit>=3.4.0,<4.0.0
pyright>=1.1.330,<2.0.0

# Extra dependencies
py-cpuinfo>=9.0.0,<10.0.0
sentence-transformers==2.2.2
pynvml>=11.5.0,<12.0.0

# Documentation
sphinx>=7.2.6,<8.0.0
mkdocs>=1.5.3,<2.0.0
mkdocs-material>=9.3.1,<10.0.0

# Optional Enhancements
pip-tools>=7.3.0,<8.0.0
loguru>=0.7.2,<1.0.0
rich>=14.0.0,<15.0.0
optimum>=1.16.0,<2.0.0
ml-dtypes>=0.3.1,<1.0.0